+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `CopyContext.exclude(Object)` and `CopyContext.include(Object)` to temporarily exclude objects from being
+  copied without registering them in the session's clone map.
+
 ## [1.0.0] - 2026-03-26
 This is the initial release of object-copier.